Rename src into bot
