+++ bot/calendars/constants/messages.py
