+++ bot/calendars/keyboards.py
@@ -1,8 +1,8 @@
 from aiogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 from aiogram.utils.keyboard import InlineKeyboardBuilder
 
-from src.core.buttons.base import ButtonGroup
-from src.core.keyboards import BaseKeyboard
+from bot.core.buttons.base import ButtonGroup
+from bot.core.keyboards import BaseKeyboard
 
 from .constants import buttons
