+++ bot/start/containers.py
@@ -2,7 +2,7 @@ from os import path
 
 from dependency_injector import containers, providers
 
-from src.core.db.utils import init_db_connection_pool, load_queries
+from bot.core.db.utils import init_db_connection_pool, load_queries
 
 from .db.queries.builders import UserQueryBuilder
 from .db.repositories import UserRepository
@@ -12,7 +12,7 @@ from .services.users import UserService
 class StartContainer(containers.DeclarativeContainer):
     wiring_config = containers.WiringConfiguration(
         modules=[
-            "src.start.routers",
+            "bot.start.routers",
         ],
     )
