+++ tests/unit/calendars/mocks/repositories.py
@@ -1,8 +1,8 @@
 from uuid import UUID, uuid4
 
-from src.calendars.constants.calendar_category import CalendarCategory
-from src.calendars.errors import CalendarDuplicateError
-from src.calendars.models.calendars import Calendar
+from bot.calendars.constants.calendar_category import CalendarCategory
+from bot.calendars.errors import CalendarDuplicateError
+from bot.calendars.models.calendars import Calendar
 
 
 class CalendarRepoositoryMock:
