+++ tests/unit/start/conftest.py
@@ -1,8 +1,8 @@
 import pytest
 
-from src.settings import Settings
-from src.start.containers import StartContainer
-from src.start.services.users import UserService
+from bot.settings import Settings
+from bot.start.containers import StartContainer
+from bot.start.services.users import UserService
 
 from .mocks.repositories import UserRepositoryMock
